requirements.txt
Django>=5.2.7
psycopg2-binary==2.9.9
python-decouple==3.8
dj-database-url==2.1.0
python-dotenv==1.0.0
djangorestframework==3.14.0
djangorestframework-simplejwt==5.3.0
spacy==3.7.2
rapidfuzz==3.5.2
scikit-learn==1.3.2
pdfplumber==0.10.3
python-docx==1.1.0
openai>=1.3.5
sentence-transformers==2.2.2
pandas==2.1.3
numpy==1.26.2
xhtml2pdf==0.2.11
requests==2.31.0
beautifulsoup4==4.12.2
celery==5.3.4
redis==5.0.1
django-cors-headers==4.3.1
whitenoise==6.6.0
supabase==2.0.3
pgvector==0.4.2
pydantic==2.5.2huggingface_hub>=0.36.0
